requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
numpy
torch
torchvision
Pillow
streamlit
streamlit
